Clean up the house
