+++ karma.config.js
@@ -4,11 +4,11 @@ module.exports = function (config) {
   config.set({
     basePath: __dirname,
     files: [
-      'test/**/*.js'
+      '*.test.js'
     ],
     frameworks: ['jasmine'],
     preprocessors: {
-      'test/**/*.js': ['webpack', 'sourcemap']
+      '*.test.js': ['webpack', 'sourcemap']
     },
     webpack: {
       module: webpackConfig.module
