+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "scrollto-with-animation",
-  "version": "4.3.3",
+  "version": "4.4.3",
   "author": "David Sancho <dsnxmoreno@gmail.com> (https://github.com/davesnx)",
   "description": "Animated Scroll with requestAnimationFrame. For smooth animate scrollTo defining the easing, running at 60FPS and cross-browser",
   "license": "MIT",
